.claude/skills/leadpilot/requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
